Rewrite + Threading
Rewrote everything just to see if i left in some unused variables, or non erroneous but weird lines of code.

CSV validation is set to multithreading. It just speeds up the validation

Downloading and Conversion are set to single thread. It's set up with the same structure to allow MT but I'm getting file accessing issues, and possibly youtube connection throttling.